{{ cookiecutter.repo_name }}/requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv==0.19.0
pandas~=1.3.2
numpy~=1.19.2
PyYAML>=5.4
scikit-learn~=0.24.0
hyperopt~=0.2.5
pytest~=6.2.2
tableone~=0.7.10
requests~=2.25.1
Keras==2.6.0

tensorflow==2.6.0
tensorflow-addons==0.14.0
pillow==8.2.0
opencv-contrib-python~=4.5.2.52
{% if cookiecutter.python_interpreter != 'python3' %}


# backwards compatibility
pathlib2
{% endif %}
